Copied everything to separate folder
I have done alot of improvements in the code readability and fixed a few
bugs. Also fixed CSS for shoppingcart notice on the right side of the
menu. I also made the menu dynamic on the clientside when user is logged
in och logged out. I have also started fixing CSS for rating and reviews
on productpages. Next step will be to code output and/or input of
reviews and ratings.